Code Test Version - Fixed a couple bugs, integrated some sfx and music.
Fixed one major bug that caused the player to not be bound within the "arena" on the Y axis. Made some adjustments to Player and Enemy controllers.
Audio has been added for player death, firing projectiles, enemy death and enemy firing projectiles. Background music was also added.
Version to be used for the 2019/6/14 deadline.